+++ javascripts/init_game.js
@@ -1,5 +1,9 @@
 /*Offline Progress*/
 function simulateTime(seconds, actual, testing) {
+	if (seconds > 10) {
+		document.getElementById("offlinePopup").style.display = "block";
+		document.getElementById("offlineProgress").innerText = "Simulating " + seconds + " seconds of progress.";
+	}
 	let ticks = seconds * 20;
 	let tickInterval = 50;
 	if (ticks > 1000 & !actual) {
@@ -8,23 +12,22 @@ function simulateTime(seconds, actual, testing) {
 	}
 	let start = Object.assign({}, player);
 	for (let complete = 0; complete < ticks; complete++) {
-		if (testing) {
-			buyMaxAll();
-		}
 		updateGame(tickInterval)
 	}
 	player.time += seconds * 1000;
-	player.meltdown.time += seconds * 1000;
 	let offlinePopup = "While you were away, "
 	if (player.energy.gt(start.energy)) {
-		offlinePopup += "your energy increased by " + notation(player.energy.log10() - start.energy.log10()) + " Orders of Magnitude.";
+		offlinePopup += "your energy increased by " + notation(player.energy.logBase(10) - start.energy.logBase(10)) + " Orders of Magnitude.";
 	}
 	if (offlinePopup == "While you were away, ") {
-		offlinePopup += "nothing happened.";
+		offlinePopup += "nothing happened...";
+	}
+	if (seconds > 1) {
+		document.getElementById("offlinePopup").style.display = "none";
 	}
 	if (seconds > 1000) {
 		document.getElementById("offlinePopup").style.display = "block";
-	    document.getElementById("offlineProgress").innerHTML = offlinePopup;
+		document.getElementById("offlineProgress").innerText = offlinePopup;
 	}
 }
 
@@ -35,7 +38,7 @@ function closeOfflineProgress() {
 function init_game() {
 	loadSave();
 	document.getElementById("production").style.display = "none";
-	document.getElementById("resources").style.display = "none";
+	document.getElementById("fuel").style.display = "none";
 	document.getElementById("mines").style.display = "none";
 	document.getElementById("reactors").style.display = "none";
 	document.getElementById("statistics").style.display = "none";
@@ -43,6 +46,7 @@ function init_game() {
 	document.getElementById("nanite").style.display = "none";
 	showNaviTab(player.navigation.naviTab);
 	targetedNotationChange(player.options.notation);
+	targetedThemeChange(player.options.theme);
 	if (Date.now() > player.lastUpdate + 1000) {
 		simulateTime((Date.now() - player.lastUpdate) / 1000);
 	}